+++ app/presenters/activity_presenter.rb
@@ -153,6 +153,20 @@ class ActivityPresenter < SimpleDelegator
     I18n.t("activity.policy_markers.#{super}")
   end
 
+  def sustainable_development_goals
+    goals = [sdg_1, sdg_2, sdg_3].compact
+    return if goals.blank?
+
+    html = "<ol class=\"govuk-list govuk-list--number\">"
+
+    goals.each do |goal|
+      html += "<li>" + I18n.t("form.label.activity.sdg_options.#{goal}") + "</li>"
+    end
+
+    html += "</ol>"
+    html.html_safe
+  end
+
   def oda_eligibility
     return if super.blank?
     I18n.t("activity.oda_eligibility.#{super}")