+++ spec/support/form_helpers.rb
@@ -35,6 +35,7 @@ module FormHelpers
     gdi: "GDI not applicable",
     collaboration_type: "Bilateral",
     flow: "ODA",
+    sdg_1: 1,
     aid_type: "B02",
     fstc_applies: true,
     policy_marker_gender: "Not assessed",
@@ -205,6 +206,14 @@ module FormHelpers
     expect(page.html).to include t("form.hint.activity.flow")
     select flow, from: "activity[flow]"
     click_button t("form.button.activity.submit")
+
+    unless level == "fund"
+      expect(page).to have_content t("form.legend.activity.sdgs_apply")
+      expect(page).to have_content t("form.hint.activity.sdgs_apply")
+      choose t("form.label.activity.sdgs_apply_options.true")
+      select t("form.label.activity.sdg_options.5"), from: "activity[sdg_1]"
+      click_button t("form.button.activity.submit")
+    end
 
     expect(page).to have_content t("form.legend.activity.aid_type")
     expect(page).to have_content t("form.hint.activity.aid_type")