+++ CHANGELOG.md
@@ -387,6 +387,7 @@
 - Add a field to report whether the activity has any relation to the Covid-19 pandemic.
 - Include the Covid-19 field in the CSV report.
 - Append the text "COVID-19" to the activity description in the IATI XML export, where applicable.
+- Collect Sustainable Development Goals (SDGs) for activities
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-22...HEAD
 [release-22]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-21...release-22